backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
Pillow==10.4.0
reportlab==4.2.2
python-dotenv==1.0.1
groq==0.9.0
tensorflow-cpu==2.16.1
Werkzeug==3.0.3
gunicorn==22.0.0
